Reorganizing routes
xrf_explorer/__init__.py:
all routes are moved to their matching file in server/routes folder

server/routes/__init__.py now ensure that these new routes are correctly imported as well
