+++ xrf_explorer/__init__.py
@@ -1,7 +1,6 @@
 # Import the necessary modules and create the Flask app
 from pathlib import Path
-
-from flask import Flask, jsonify, request, send_from_directory
+from flask import Flask, send_from_directory
 from flask_cors import CORS
 from flask_httpauth import HTTPTokenAuth
 from xrf_explorer.server.database.auth import provide_auth
@@ -9,12 +8,19 @@ from xrf_explorer.server.database.database import init_app
 from xrf_explorer.server.database.models import User  # Ensure models are imported
 from xrf_explorer.server.database.models import UserRole
 
+# Create the Flask app
 app: Flask = Flask(__name__, template_folder=Path('client/templates'), static_folder='client/dist')
+
+# Enable CORS for the app
 CORS(app)
+
+# Initialize the database
 db = init_app(app)
 
+# Set up authentication
 auth: HTTPTokenAuth = provide_auth()
 
+# Create the database tables and a default admin user if none exist
 with app.app_context():
     db.create_all()  # Create database tables for our data models
     
@@ -38,103 +44,8 @@ with app.app_context():
     else:
         print("Viewer user already exists.")
 
+# Import and register the routes
 from xrf_explorer.server.routes import *
-
-# Login route
-@app.route('/api/login', methods=['POST'])
-def login():
-    
-    # Get the username and password from the request
-    data = request.json
-    username = data.get('username')
-    password = data.get('password')
-    
-    # Check if the username and password are provided
-    if not username or not password:
-        return jsonify({"success": False, "message": "Missing username or password"}), 400
-
-    # Find the user in the database
-    user = User.query.filter_by(username=username).first()
-    
-    # Verify the password
-    if user and user.check_password(password):
-        # token = user.generate_auth_token()
-
-        # return jsonify({"success": True, "message": "Login successful", "role": user.role.name, "token": token}), 200
-        return jsonify({"success": True, "message": "Login successful", "role": user.role.name}), 200
-    return jsonify({"success": False, "message": "Invalid username or password"}), 401
-
-# Account creation route
-@auth.login_required(role=UserRole.ADMIN)
-@app.route('/api/create_account', methods=['POST'])
-def create_account():
-    data = request.json
-    username = data.get('username')
-    password = data.get('password')
-    role_str = data.get('role')
-
-    # Check if the username, password, and role are provided
-    if not username or not password or not role_str:
-        return jsonify({"success": False, "message": "Missing username, password, or role"}), 400
-
-    # Validate the role
-    try:
-        role = UserRole[role_str]
-    except KeyError:
-        return jsonify({"success": False, "message": "Invalid role"}), 400
-
-    # Check if the username already exists
-    if User.query.filter_by(username=username).first():
-        return jsonify({"success": False, "message": "Username already exists"}), 409
-
-    # Create a new user
-    new_user = User(username=username, role=role)
-    new_user.set_password(password)
-    db.session.add(new_user)
-    db.session.commit()
-
-    return jsonify({"success": True, "message": "Account created successfully"}), 201
-
-# Account retrieval route
-@auth.login_required(role=UserRole.ADMIN)
-@app.route('/api/accounts', methods=['GET'])
-def get_accounts():
-    accounts = User.query.all()
-    return jsonify([{"username": acc.username, "role": acc.role.name.capitalize()} for acc in accounts]), 200
-
-# Account update route
-@auth.login_required(role=UserRole.ADMIN)
-@app.route('/api/update_account', methods=['POST'])
-def update_account():
-    data = request.json
-    originalUsername = data.get('originalUsername')
-    username = data.get('username')
-    password = data.get('password')
-    role_str = data.get('role')
-
-    # Check if the username and role are provided
-    if not username or not role_str:
-        return jsonify({"success": False, "message": "Missing username or role"}), 400
-
-    # Validate the role
-    try:
-        role = UserRole[role_str]
-    except KeyError:
-        return jsonify({"success": False, "message": "Invalid role"}), 400
-
-    # Check if the user is already in the database
-    if User.query.filter_by(username=username).first() and (username != originalUsername):
-        return jsonify({"success": False, "message": "Username already exists"}), 409
-
-    # Update the user's username, password and role
-    user = User.query.filter_by(username=originalUsername).first()
-    user.username = username
-    user.role = role
-    if (password): 
-        user.set_password(password)
-    db.session.commit()
-
-    return jsonify({"success": True, "message": "Account updated successfully"}), 200
 
 # All routes not matched in the server are forwarded to the client
 @app.route('/', defaults={'path': 'index.html'})
